requirements.txt
langchain>=0.1.0
langchain-community>=0.0.30
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
python-dotenv>=1.0.0
pdfplumber>=0.9.0
pydantic>=2.6.1
tqdm>=4.66.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
pytest>=7.4.0
